pyproject.toml
[tool.poetry]
name = "ai_assistant_telegram_bot"

version = "0.0.1"
description = ""
authors = ["Andrey Khaykin"]

[tool.poetry.dependencies]
python = "~3.10"
flask = "^2.2.3"
pytelegrambotapi = "^4.10.0"
python-dotenv = "^1.0.0"
pillow = "^9.4.0"
torch = "^1.13.1"
opencv-python = "^4.7.0.72"
openai-whisper = "^20230124"
openai = "^0.27.0"
tts = "^0.11.1"
setuptools-rust = "^1.5.2"
omegaconf = "^2.3.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"
black = "22.3.0"
pyupdate = "^1.4.0"
pyupgrade = "^1.1.0"
isort = "^5.6.4"
flake8 = "^3.8.4"


[tool.black]
# https://github.com/psf/black
line-length = 88
target-version = ["py37"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
include_trailing_comma = true
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
